Make Cypress be a little more patient
I'm not sure the number of todos you actually expect here, but I raised the number by 1 so it's going to wait just a little longer before trying to do the next thing. The `cy.get` action is going to hold things up until we have what's there or we time out. This will make detaching from the DOM less likely.